read hash file

+++ cmd/es-devnet/test.sh
@@ -16,21 +16,24 @@ executable="./es-devnet"
 data_dir="./es-data"
 storage_file_0="$data_dir/shard-0.dat"
 
-# remove old file
-if [ $storage_file_0 ]; then
-  rm -r $data_dir
-  echo "remove ${storage_file_0}"
+generate_data="false"
+if [ "$REMOVE_FILES" == "true" ]; then
+  generate_data="true"
+  # remove old file
+  if [ $storage_file_0 ]; then
+    rm -r $data_dir
+    echo "remove ${data_dir}"
+  fi
 fi
 
-
-# init shard 0
-init_shard=" --datadir $data_dir \
+# create data and upload to contract
+generate_test_data=" --datadir $data_dir \
   --l1.rpc http://65.108.236.27:8545 \
   --l1.chainId 7011893059 \
   --storage.l1contract 0x7Ff73D2eE68ccB19FB5F62e210D8Db5E98dD7F1e \
   --storage.privateKey $ES_NODE_PRIVATE_KEY \
   --storage.miner $ES_NODE_STORAGE_MINER \
+  --generateData $generate_data \
   --shardLength 1
   "
-
-$executable $init_shard
+$executable $generate_test_data
